Create account feature.
The user can now create an account and also create an order. Needs to be connected to the dB!
