+++ discord/discordbot.py
@@ -1,12 +1,10 @@
 import discord
 import discordRequestHandler
-
+import add
 client = discord.Client()
-
 
 @client.event
 async def on_ready():
-    await client.channel.send("Hi, I am Bytes the Pink Pythons team bot. For a list of commands type '!help'.")
     print("Connecting as", client.user.name)
 
 
@@ -15,8 +13,10 @@ async def on_message(message):
     # Check if the user who sent the message is this bot.
     if message.author == client.user:
         return
-    await discordRequestHandler.handle(message.upper())
-
+    #Check if message starts with !
+    if message.content.startswith("!"):
+    #Next check if the user already has an account
+        await discordRequestHandler.handle(message)
 
 # I dont want to store a secret api key on discord
 # Dont upload the code.txt file!!!
